feat(ftso): signed getCurrentFeed read family
Change the unpublished per-feed-timestamp reads to signed (int256)
values and complete them into a full family — getCurrentFeed and
getCurrentFeedInWei singles alongside the getCurrentFeeds batches — so
a custom feed with a signed source can report negative values through
FtsoV2. IICustomFeed.getCurrentFeed returns int256: the selector and
the return ABI encoding are unchanged for non-negative values, so
custom feeds deployed against the unsigned declaration remain
compatible; SFlrCustomFeed and StXrpCustomFeed sources get
reinterpreting casts with safety comments.

The published unsigned read paths (getFeedsById family) reject
negative custom feed values with "value negative" — documented on
each of them with a pointer to the signed variant; the signed wei
conversion truncates negative values toward zero when scaling down.
Feed-id rename resolution is consolidated in the shared internal
single-read used by both signed and unsigned paths.

Includes tests for the new reads and the negative-value matrix,
documentation updates and a changelog entry.

+++ contracts/customFeeds/implementation/StXrpCustomFeed.sol
@@ -45,7 +45,7 @@ contract StXrpCustomFeed is IICustomFeed {
     /**
      * @inheritdoc IICustomFeed
      */
-    function getCurrentFeed() external payable returns (uint256 _value, int8 _decimals, uint64 _timestamp) {
+    function getCurrentFeed() external payable returns (int256 _value, int8 _decimals, uint64 _timestamp) {
         IFastUpdatesConfiguration fastUpdatesConfiguration =
             IFastUpdatesConfiguration(flareContractRegistry.getContractAddressByName("FastUpdatesConfiguration"));
         IFastUpdater fastUpdater = IFastUpdater(flareContractRegistry.getContractAddressByName("FastUpdater"));
@@ -54,7 +54,10 @@ contract StXrpCustomFeed is IICustomFeed {
         indices[0] = fastUpdatesConfiguration.getFeedIndex(referenceFeedId);
         (uint256[] memory values, int8[] memory decimals, uint64 timestamp) =
             fastUpdater.fetchCurrentFeeds{value: msg.value} (indices);
-        _value = stXrp.convertToAssets(values[0]);
+        // IICustomFeed is signed for feeds with signed sources; this feed's value is inherently
+        // non-negative, and an asset conversion of an FTSO price sits far below 2^255, so the
+        // reinterpreting cast cannot wrap into a negative number.
+        _value = int256(stXrp.convertToAssets(values[0]));
         _decimals = decimals[0];
         _timestamp = timestamp;
     }
